If we're going to be passing a negative value for the size of a stringfield, in order to indicate something, then using an UNSIGNED parameter is bad, mmmmmkay?
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@89312 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ include/asterisk/stringfields.h
@@ -230,7 +230,7 @@ void __ast_string_field_ptr_build_va(struct ast_string_field_mgr *mgr,
 
 /*! \internal \brief internal version of ast_string_field_init */
 int __ast_string_field_init(struct ast_string_field_mgr *mgr,
-	struct ast_string_field_pool **pool_head, size_t needed);
+	struct ast_string_field_pool **pool_head, int needed);
 
 /*!
   \brief Set a field to a simple string value
